refactor: 각 커스텀 훅에서 훅과 직접 연관되지 않은 함수를 util.ts로 분리
