+++ hooks/src/lib/useCardBrand/useCardBrand.ts
@@ -1,7 +1,8 @@
 import { useState } from 'react';
 import ValidationResult from '../types/ValidationResult';
-
 import DEFAULT_CARD_BRANDS from '../constants/defaultCardBrands';
+
+import { getValidationResult, validateInitialParams } from './useCardBrand.util';
 
 interface CardBrandParams {
   allowedBrands?: string[];
@@ -15,7 +16,7 @@ interface BrandValidationResult {
   handleUpdateBrand: (value: string) => void;
 }
 
-interface CardBrandErrorMessages {
+export interface CardBrandErrorMessages {
   invalidBrand: string;
   emptyAllowedBrands: string;
   initialValueNotExistsInAllowedBrands: string;
@@ -51,41 +52,4 @@ export default function useCardBrand({
   };
 
   return { brand, validationResult, handleUpdateBrand };
-}
-
-function validateInitialParams(
-  allowedBrands: string[],
-  initialValue: string,
-  errorMessages: CardBrandErrorMessages,
-) {
-  if (allowedBrands.length < 1) {
-    throw new Error(errorMessages.emptyAllowedBrands);
-  }
-
-  if (!validateBrand([...allowedBrands, ''], initialValue)) {
-    throw new Error(errorMessages.initialValueNotExistsInAllowedBrands);
-  }
-}
-
-function getValidationResult(
-  allowedBrands: string[],
-  value: string,
-  errorMessages: CardBrandErrorMessages,
-) {
-  if (value === DEFAULT_PARAMS.initialValue) {
-    return { isValid: null };
-  }
-
-  if (!validateBrand(allowedBrands, value)) {
-    return {
-      isValid: false,
-      errorMessage: errorMessages.invalidBrand,
-    };
-  }
-
-  return { isValid: true };
-}
-
-function validateBrand(allowedBrands: string[], value: string) {
-  return allowedBrands.includes(value);
 }
