+++ hooks/src/lib/useCardCVC/useCardCVC.ts
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import ValidationResult from '../types/ValidationResult';
-import Validation from '../utils/Validation';
 import ErrorMessages from '../types/ErrorMessages';
 
 import { validateAllowedLength } from '../utils/validateInitialParams';
+import { getValidationResult } from './useCardCVC.util';
 
 interface CardCVCParams {
   allowedLength?: number;
@@ -59,20 +59,4 @@ export default function useCardCVC({
   };
 
   return { CVC, validationResult, handleUpdateCVC };
-}
-
-function getValidationResult(value: string, allowedLength: number, errorMessages: ErrorMessages) {
-  if (value === DEFAULT_PARAMS.initialValue) {
-    return { isValid: null };
-  }
-
-  if (!Validation.isNumeric(value)) {
-    return { isValid: false, errorMessage: errorMessages.inputType };
-  }
-
-  if (!Validation.hasLength(value, allowedLength)) {
-    return { isValid: false, errorMessage: errorMessages.inputLength!(allowedLength) };
-  }
-
-  return { isValid: true };
 }
